LayoutTests/imported/w3c/web-platform-tests/fetch/api/response/response-error.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Response error</title>
    <meta name="help" href="https://fetch.spec.whatwg.org/#response">
    <meta name="author" title="Canon Research France" href="https://www.crf.canon.fr">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      test(function() {
        var invalidStatus = [0, 100, 199, 600, 1000];
        invalidStatus.forEach(function(status) {
          assert_throws(new RangeError() , function() { new Response("", { "status" : status }); },
            "Expect RangeError exception when status is " + status);
        });
      },"Throws RangeError when responseInit's status is invalid");

      test(function() {
        var invalidStatusText = ["\n", "Ā"];
        invalidStatusText.forEach(function(statusText) {
          assert_throws(new TypeError() , function() { new Response("", { "statusText" : statusText }); },
            "Expect TypeError exception " + statusText);
        });
      },"Throws TypeError when responseInit's statusText is invalid");

      test(function() {
        var nullBodyStatus = [204, 205, 304];
        nullBodyStatus.forEach(function(status) {
          assert_throws(new TypeError() ,
            function() { new Response("body", { "status" : status }); },
            "Expect TypeError exception ");
        });
      },"Throws TypeError when building a response with body and a null body status");
    </script>
  </body>
</html>
